Cleaned up and restructured behavior
