+++ src/man/noggin/players/PositionStates.py
@@ -57,148 +57,3 @@ def relocalize(player):
 
     return player.stay()
 
-def afterPenalty(player):
-#Management State
-
-    if player.lastDiffState == 'penaltyLookLeft':
-        return player.goLater('penaltyLookRight')
-
-    if player.lastDiffState == 'penaltyLookRight':
-        return player.goLater('penaltyRelocalize')
-
-    return player.goLater('penaltyLookLeft')
-
-def penaltyLookLeft(player):
-
-    if player.firstFrame():
-        player.headCount = 0
-        player.yellowCount = 0
-        player.blueCount = 0
-        player.ballCount = 0
-        player.brain.tracker.performHeadMove(HeadMoves.LOOK_UP_LEFT)
-
-    if player.brain.ball.on:
-        player.ballCount +=1
-        if player.ballCount == 3:
-            #deal with ball and don't worry about loc
-            player.brain.tracker.trackBall()
-            return player.goLater('gamePlaying')
-
-    if not player.brain.motion.isHeadActive():
-        ##looking left
-
-        if player.brain.yglp.on or player.brain.ygrp.on:
-            #see the goal posts in multiple frames for safety
-            player.blueCount = 0
-            player.yellowCount += 1
-            if player.yellowCount == 5:
-                #set loc info
-                player.brain.loc.resetLocTo(NogginConstants.CENTER_FIELD_X, \
-                                        NogginConstants.FIELD_WHITE_TOP_SIDELINE_Y, \
-                                        -90.0)
-                #now you know where you are!
-                return player.goLater('gamePlaying')
-
-        if player.brain.bglp.on or player.brain.bgrp.on:
-            #see the goal posts in multiple frames for safety
-            player.yellowCount = 0
-            player.blueCount += 1
-            if player.blueCount == 5:
-                #set loc info
-                player.brain.loc.resetLocTo(NogginConstants.CENTER_FIELD_X, \
-                                        NogginConstants.FIELD_WHITE_BOTTOM_SIDELINE_Y, \
-                                        90.0)
-                #now you know where you are!
-                return player.goLater('gamePlaying')
-
-        player.headCount += 1
-
-    #if we are looking for too long
-    if player.headCount > 10:
-        return player.goLater(player.lastDiffState)
-
-    return player.stay()
-
-def penaltyLookRight(player):
-
-    if player.firstFrame():
-        player.headCount = 0
-        player.yellowCount = 0
-        player.blueCount = 0
-        player.ballCount = 0
-        player.brain.tracker.performHeadMove(HeadMoves.LOOK_UP_RIGHT)
-
-
-    if player.brain.ball.on:
-        player.ballCount += 1
-        if player.ballCount == 3:
-            #deal with ball and don't worry about loc
-            player.brain.tracker.trackBall()
-            return player.goLater('gamePlaying')
-
-    if not player.brain.motion.isHeadActive():
-        ##looking right
-
-        if player.brain.yglp.on or player.brain.ygrp.on:
-            #see the goal posts in multiple frames for safety
-            player.blueCount = 0
-            player.yellowCount += 1
-            if player.yellowCount == 5:
-                #set loc info
-                player.brain.loc.resetLocTo(NogginConstants.CENTER_FIELD_X, \
-                                        NogginConstants.FIELD_WHITE_BOTTOM_SIDELINE_Y, \
-                                        90.0)
-                #now you know where you are!
-                return player.goLater('gamePlaying')
-
-        if player.brain.bglp.on or player.brain.bgrp.on:
-            #see the goal posts in multiple frames for safety
-            player.yellowCount = 0
-            player.blueCount += 1
-            if player.blueCount == 5:
-                #set loc info
-                player.brain.loc.resetLocTo(NogginConstants.CENTER_FIELD_X, \
-                                        NogginConstants.FIELD_WHITE_TOP_SIDELINE_Y, \
-                                        -90.0)
-                #now you know where you are!
-                return player.goLater('gamePlaying')
-
-        player.headCount += 1
-
-    #if we are looking for too long
-    if player.headCount > 10:
-        #no luck
-        return player.goLater(player.lastDiffState)
-
-    return player.stay()
-
-def penaltyRelocalize(player):
-    if player.firstFrame():
-        player.setWalk(10 , 0, 0)
-
-    if player.brain.ball.on:
-        player.brain.tracker.trackBall()
-        return player.goLater('gamePlaying')
-
-    if player.brain.my.locScore == NogginConstants.GOOD_LOC or \
-            player.brain.my.locScore == NogginConstants.OK_LOC:
-        player.shouldRelocalizeCounter += 1
-
-        if player.shouldRelocalizeCounter > 30:
-            player.shouldRelocalizeCounter = 0
-            return player.goLater('gamePlaying')
-
-    else:
-        player.shouldRelocalizeCounter = 0
-
-    if not player.brain.motion.isHeadActive():
-        player.brain.tracker.locPans()
-
-    if player.counter > constants.RELOC_SPIN_FRAME_THRESH:
-        direction = MyMath.sign(player.getWalk()[2])
-        if direction == 0:
-            direction = 1
-
-        player.setWalk(0 , 0, constants.RELOC_SPIN_SPEED * direction)
-
-    return player.stay()
